requirements.txt
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.0
idna==3.10
mysql-connector-python==9.1.0
prettytable==3.12.0
requests==2.32.3
soupsieve==2.6
urllib3==2.2.3
wcwidth==0.2.13
